Adds resources and refactor HtmlOutputUtil.java
