+++ main/src/main/resources/script.js
@@ -1,0 +1,5 @@
+$(function(){
+  $(".accordion tr.notice").on("click", function(){
+    $(this).toggleClass("open").next(".description").toggleClass("open");
+  });
+});
